+++ packages/dashmate/src/commands/setup-for-local-development.js
@@ -6,6 +6,7 @@ const BaseCommand = require('../oclif/command/BaseCommand');
 const MuteOneLineError = require('../oclif/errors/MuteOneLineError');
 
 const NETWORKS = require('../networks');
+const wait = require('../util/wait');
 
 class SetupForLocalDevelopmentCommand extends BaseCommand {
   /**
@@ -64,6 +65,12 @@ class SetupForLocalDevelopmentCommand extends BaseCommand {
                   isMinerEnabled: true,
                 },
               ),
+            },
+            {
+              title: 'Wait 20 seconds to ensure all services are running',
+              task: async () => {
+                await wait(20000);
+              },
             },
             {
               title: 'Initialize Platform',