+++ packages/dashmate/src/listr/tasks/startNodeTaskFactory.js
@@ -4,8 +4,6 @@ const { PrivateKey } = require('@dashevo/dashcore-lib');
 
 const NETWORKS = require('../../networks');
 
-const wait = require('../../util/wait');
-
 /**
  *
  * @param {DockerCompose} dockerCompose
@@ -16,7 +14,6 @@ function startNodeTaskFactory(dockerCompose) {
    * @typedef {startNodeTask}
    * @param {Config} config
    * @param {Object} options
-   * @param {boolean} [options.isFullNode]
    * @param {string} [options.driveImageBuildPath]
    * @param {string} [options.dapiImageBuildPath]
    * @param {boolean} [options.isUpdate]
@@ -26,7 +23,6 @@ function startNodeTaskFactory(dockerCompose) {
   function startNodeTask(
     config,
     {
-      isFullNode,
       driveImageBuildPath = undefined,
       dapiImageBuildPath = undefined,
       isUpdate = undefined,
@@ -36,8 +32,13 @@ function startNodeTaskFactory(dockerCompose) {
     // Check external IP is set
     config.get('externalIp', true);
 
+    if (isMinerEnabled === undefined) {
+      // eslint-disable-next-line no-param-reassign
+      isMinerEnabled = config.get('core.miner.enable');
+    }
+
     if (isMinerEnabled === true && config.get('network') !== NETWORKS.LOCAL) {
-      this.error(`'core.miner.enabled' option supposed to work only with local network. Your network is ${config.get('network')}`, { exit: true });
+      throw new Error(`'core.miner.enabled' option only works with local network. Your network is ${config.get('network')}.`);
     }
 
     return new Listr([
@@ -47,9 +48,18 @@ function startNodeTaskFactory(dockerCompose) {
         task: async () => dockerCompose.pull(config.toEnvs()),
       },
       {
+        title: 'Check node is not started',
+        task: async () => {
+          if (await dockerCompose.isServiceRunning(config.toEnvs())) {
+            throw new Error('Running services detected. Please ensure all services are stopped for this config before starting');
+          }
+        },
+      },
+      {
         title: 'Start services',
         task: async () => {
-          if (!isFullNode) {
+          const isMasternode = config.get('core.masternode.enable');
+          if (isMasternode) {
             // Check operatorPrivateKey is set
             config.get('core.masternode.operator.privateKey', true);
           }
@@ -58,7 +68,7 @@ function startNodeTaskFactory(dockerCompose) {
 
           if (driveImageBuildPath || dapiImageBuildPath) {
             if (config.get('network') === NETWORKS.TESTNET) {
-              this.error('You can\'t use drive-image-build-path and dapi-image-build-path options with testnet network', { exit: true });
+              throw new Error('You can\'t use drive-image-build-path and dapi-image-build-path options with testnet network');
             }
 
             if (driveImageBuildPath) {
@@ -73,9 +83,6 @@ function startNodeTaskFactory(dockerCompose) {
           }
 
           await dockerCompose.up(envs);
-
-          // wait 10 seconds to ensure all services are running
-          await wait(10000);
         },
       },
       {